Creates board, square, and piece classes
The Board is a 2D array of Squares, and each Square can have a Piece on it.  The
Square and the Piece have textures assigned to them as well as colors and piece
types.  The Board itself is in charge of creating the grid of Squares, and
managing Piece movement.